Fix current fence value being equal to initial fence value.
With the current fence value being set to the same value as the initial fence
values, it was impossible to identify if the first fence was hit because the
fence value was set to 0 or because the current fence value was 0 at the
beginning. Initialize to 1 so on the first reuse, we'll correctly wait on the
first fence.